# docs/swagger/swagger.yaml
definitions:
...
  ErrorResponse:
    properties:
      code:
        example: VALIDATION_ERROR
        type: string
      details: {}
      error:
        example: Invalid request
        type: string
      success:
        example: false
        type: boolean
    type: object
...
  SuccessResponse:
    properties:
      data: {}
      message:
        example: Operation completed successfully
        type: string
      success:
        example: true
        type: boolean
    type: object
...
  zpwoot_internal_services_shared_dto.CreateGroupRequest:
    properties:
      description:
        maxLength: 512
        type: string
      name:
        maxLength: 25
        minLength: 1
        type: string
      participants:
        items:
          type: string
        maxItems: 256
        minItems: 1
        type: array
    required:
    - name
    - participants
    type: object
  zpwoot_internal_services_shared_dto.CreateGroupResponse:
    properties:
      created_at:
        type: string
      description:
        type: string
      group_jid:
        type: string
      message:
        type: string
      name:
        type: string
      participants:
        items:
          type: string
        type: array
      success:
        type: boolean
    type: object
  zpwoot_internal_services_shared_dto.GetGroupInfoResponse:
    properties:
      created_at:
        type: string
      description:
        type: string
      group_jid:
        type: string
      message:
        type: string
      name:
        type: string
      owner:
        type: string
      participants:
        items:
          $ref: '#/definitions/zpwoot_internal_services_shared_dto.ParticipantInfo'
        type: array
      settings:
        $ref: '#/definitions/zpwoot_internal_services_shared_dto.GroupSettings'
      success:
        type: boolean
      updated_at:
        type: string
    type: object
  zpwoot_internal_services_shared_dto.GroupInfo:
    properties:
      created_at:
        type: string
      description:
        type: string
      group_jid:
        type: string
      name:
        type: string
      owner:
        type: string
      participants:
        type: integer
    type: object
  zpwoot_internal_services_shared_dto.GroupSettings:
    properties:
      announce:
        type: boolean
      join_approval_mode:
        type: string
      locked:
        type: boolean
      member_add_mode:
        type: string
      restrict:
        type: boolean
    type: object
  zpwoot_internal_services_shared_dto.ListGroupsResponse:
    properties:
      count:
        type: integer
      groups:
        items:
          $ref: '#/definitions/zpwoot_internal_services_shared_dto.GroupInfo'
        type: array
      message:
        type: string
      success:
        type: boolean
    type: object
...
  zpwoot_internal_services_shared_dto.ParticipantInfo:
    properties:
      jid:
        type: string
      joined_at:
        type: string
      role:
        type: string
      status:
        type: string
    type: object
  zpwoot_internal_services_shared_dto.SetGroupNameRequest:
    properties:
      group_jid:
        type: string
      name:
        maxLength: 25
        minLength: 1
        type: string
    required:
    - group_jid
    - name
    type: object
  zpwoot_internal_services_shared_dto.SetGroupNameResponse:
    properties:
      group_jid:
        type: string
      message:
        type: string
      name:
        type: string
      success:
        type: boolean
    type: object
  zpwoot_internal_services_shared_dto.UpdateParticipantsRequest:
    properties:
      action:
        enum:
        - add
        - remove
        - promote
        - demote
        type: string
      group_jid:
        type: string
      participants:
        items:
          type: string
        minItems: 1
        type: array
    required:
    - action
    - group_jid
    - participants
    type: object
  zpwoot_internal_services_shared_dto.UpdateParticipantsResponse:
    properties:
      action:
        type: string
      group_jid:
        type: string
      message:
        type: string
      participants:
        items:
          type: string
        type: array
      success:
        type: boolean
    type: object
info:
  contact: {}
paths:
...
  /sessions/{sessionId}/groups:
    get:
      description: List all WhatsApp groups for a session
      parameters:
      - description: Session ID
        in: path
        name: sessionId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/zpwoot_internal_services_shared_dto.ListGroupsResponse'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: List WhatsApp groups
      tags:
      - Groups
    post:
      consumes:
      - application/json
      description: Create a new WhatsApp group with specified participants
      parameters:
      - description: Session ID
        in: path
        name: sessionId
        required: true
        type: string
      - description: Group creation request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/zpwoot_internal_services_shared_dto.CreateGroupRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/zpwoot_internal_services_shared_dto.CreateGroupResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Create new WhatsApp group
      tags:
      - Groups
  /sessions/{sessionId}/groups/info:
    get:
      description: Get detailed information about a WhatsApp group
      parameters:
      - description: Session ID
        in: path
        name: sessionId
        required: true
        type: string
      - description: Group JID
        in: query
        name: groupJid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/zpwoot_internal_services_shared_dto.GetGroupInfoResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get group information
      tags:
      - Groups
  /sessions/{sessionId}/groups/name:
    put:
      consumes:
      - application/json
      description: Change the name of a WhatsApp group
      parameters:
      - description: Session ID
        in: path
        name: sessionId
        required: true
        type: string
      - description: Group name request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/zpwoot_internal_services_shared_dto.SetGroupNameRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/zpwoot_internal_services_shared_dto.SetGroupNameResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Set group name
      tags:
      - Groups
  /sessions/{sessionId}/groups/participants:
    post:
      consumes:
      - application/json
      description: Add, remove, promote or demote group participants
      parameters:
      - description: Session ID
        in: path
        name: sessionId
        required: true
        type: string
      - description: Participants update request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/zpwoot_internal_services_shared_dto.UpdateParticipantsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/zpwoot_internal_services_shared_dto.UpdateParticipantsResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Update group participants
      tags:
      - Groups
...
swagger: "2.0"